Use faster sysretq when returning from system calls

See merge request redox-os/kernel!168

+++ src/context/signal.rs
@@ -122,7 +122,7 @@ pub extern "C" fn signal_handler(sig: usize) {
             sp -= mem::size_of::<usize>();
             *(sp as *mut usize) = restorer;
 
-            usermode(handler, sp, sig, singlestep);
+            usermode(handler, sp, sig, u32::from(singlestep));
         }
     }
